requirements.txt
google-cloud-aiplatform>=1.38.0
python-dotenv>=1.0.0
Pillow>=10.0.0
matplotlib>=3.7.0
requests>=2.31.0 